requirements2.txt
astroid==2.11.6
attrs==21.4.0
autoflake==1.4
aws-cdk-lib==2.27.0
aws-cdk.aws-lambda-python-alpha==2.27.0a0
aws-lambda-types==0.2.1
bandit==1.7.4
black==22.3.0
CacheControl==0.12.11
cattrs==22.1.0
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
constructs==10.1.37
coverage==6.4.1
cyclonedx-python-lib==2.5.1
dill==0.3.5.1
distlib==0.3.4
exceptiongroup==1.0.0rc8
filelock==3.7.1
flake8==4.0.1
flake8-plugin-utils==1.3.2
flake8-pytest-style==1.6.0
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
html5lib==1.1
identify==2.5.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
isort==5.10.1
jsii==1.60.1
junit-xml==1.8
lazy-object-proxy==1.7.1
lockfile==0.12.2
mccabe==0.7.0
msgpack==1.0.4
mutmut==2.4.0
mypy==0.961
mypy-extensions==0.4.3
nodeenv==1.6.0
packageurl-python==0.9.9
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pip-api==0.0.29
pip_audit==2.3.1
platformdirs==2.5.2
pluggy==1.0.0
pony==0.7.16
pre-commit==2.19.0
progress==1.6
publication==0.0.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.14.1
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-gitignore==1.3
python-dateutil==2.8.2
pyupgrade==2.34.0
PyYAML==6.0
requests==2.28.0
resolvelib==0.8.1
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
stevedore==3.5.0
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typed-ast==1.5.4
typing-extensions==3.10.0.2
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
